+++ infrastruktur/src/main/kotlin/com/github/haschi/haushaltsbuch/infrastruktur/rest/WaehrungsbetragSerializer.kt
@@ -6,16 +6,15 @@ import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.github.haschi.dominium.haushaltsbuch.core.model.values.Währungsbetrag
 import java.io.IOException
 
-class WährungsbetragSerialisierer @JvmOverloads constructor(t: Class<Währungsbetrag>? = null)
+class WaehrungsbetragSerializer @JvmOverloads constructor(t: Class<Währungsbetrag>? = null)
     : StdSerializer<Währungsbetrag>(t)
 {
-
     @Throws(IOException::class)
     override fun serialize(
             value: Währungsbetrag,
-            jgen: JsonGenerator,
+            generator: JsonGenerator,
             provider: SerializerProvider)
     {
-        jgen.writeString(value.toString())
+        generator.writeString(value.toString())
     }
 }
