+++ ui/backend/src/main/kotlin/com/github/haschi/haushaltsbuch/backend/rest/api/InventarMixin.kt
@@ -1,5 +1,6 @@
-package com.github.haschi.haushaltsbuch.infrastruktur.rest
+package com.github.haschi.haushaltsbuch.backend.rest.api
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 
-@JsonIgnoreProperties(value = ["reinvermoegen"], allowGetters = true) abstract class InventarMixin
+@JsonIgnoreProperties(value = ["reinvermoegen"], allowGetters = true)
+abstract class InventarMixin
