+++ ui/backend/src/main/kotlin/com/github/haschi/haushaltsbuch/backend/rest/api/RestCommandMessage.kt
@@ -2,6 +2,5 @@ package com.github.haschi.haushaltsbuch.backend.rest.api
 
 data class RestCommandMessage(
         val type: String,
-        // @com.fasterxml.jackson.annotation.JsonValue
         val payload: Any,
         val meta: Map<String, String> = emptyMap())
