+++ ui/backend/src/main/kotlin/com/github/haschi/haushaltsbuch/backend/rest/api/VersionEndpoint.kt
@@ -1,7 +1,6 @@
 package com.github.haschi.haushaltsbuch.backend.rest.api
 
 import org.axonframework.commandhandling.gateway.CommandGateway
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestMethod
@@ -14,6 +13,6 @@ class VersionEndpoint(val gateway: CommandGateway)
     @ResponseBody
     fun get(): String
     {
-        return "1.0";
+        return "1.0"
     }
 }
