+++ ui/backend/src/main/kotlin/com/github/haschi/haushaltsbuch/backend/rest/api/WaehrungsbetragDeserializer.kt
@@ -1,4 +1,4 @@
-package com.github.haschi.haushaltsbuch.infrastruktur.rest
+package com.github.haschi.haushaltsbuch.backend.rest.api
 
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.DeserializationContext
@@ -6,12 +6,9 @@ import com.fasterxml.jackson.databind.deser.std.StdScalarDeserializer
 import com.github.haschi.dominium.haushaltsbuch.core.model.values.Währungsbetrag
 import java.io.IOException
 
-class WährungsbetragDeserialisierer : StdScalarDeserializer<Währungsbetrag>
+class WaehrungsbetragDeserializer
+    : StdScalarDeserializer<Währungsbetrag>(Währungsbetrag::class.java)
 {
-    constructor() : super(Währungsbetrag::class.java)
-
-    constructor(vc: Class<*>) : super(vc)
-
     @Throws(IOException::class)
     override fun deserialize(
             jsonParser: JsonParser,
