+++ ui/backend/src/main/kotlin/com/github/haschi/haushaltsbuch/backend/rest/api/WaehrungsbetragSerializer.kt
@@ -1,4 +1,4 @@
-package com.github.haschi.haushaltsbuch.infrastruktur.rest
+package com.github.haschi.haushaltsbuch.backend.rest.api
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.SerializerProvider
@@ -6,16 +6,16 @@ import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.github.haschi.dominium.haushaltsbuch.core.model.values.Währungsbetrag
 import java.io.IOException
 
-class WährungsbetragSerialisierer @JvmOverloads constructor(t: Class<Währungsbetrag>? = null)
+class WaehrungsbetragSerializer @JvmOverloads constructor(t: Class<Währungsbetrag>? = null)
     : StdSerializer<Währungsbetrag>(t)
 {
 
     @Throws(IOException::class)
     override fun serialize(
             value: Währungsbetrag,
-            jgen: JsonGenerator,
+            generator: JsonGenerator,
             provider: SerializerProvider)
     {
-        jgen.writeString(value.toString())
+        generator.writeString(value.toString())
     }
 }
